WIP: Try to avoid the context switch when the SerialDispatchQueue is idle
Use case is for something like qpid proton-hawtdispatch integration where
the proton engine is serialized with a SerialDispatchQueue and the
calling thread wants to opportunistically synchronously send an amqp
message which has to happen in the context of the SerialDispatchQueue.


Note: the current approach is not sufficient.

Under load that includes network activity the calling thread starts
to get treated as hawtdispatch thread pool member and since it doesnt
behave like that, it leads to a deadlock.

current guess is that this is caused by selector load balancing.